refactor: [US-054] - Extract dna/persistence.py (_save_dna_to_profile, save_anonymous_session_data)
